Added the ability to set the frame rate
Updated to the current master of PS3EYEDriver and added the ability to set the frame rate of
each camera. The frame rates are now stored in the configuration files and can be configured
using PSMoveConfigTool (changed the UI window sizes accordingly). test_camera also has the
ability to change the frame rate of the opened cameras with the y and h keys and check the
current setting with the space bar. The initialisation frame rate was changed to 15 fps from
60 fps to avoid bandwidth issues at the start.