Added a parallel version of test_camera
The cameras are now operated in parallel so that they can have
completely independent frame rates (entered by the user at the
start). Previously it was limited by the slowest rate. All other settings
are also unique to each camera and can be set by clicking on it's window
and a list of the controlls is given after entering the frame rates.
The camera limit has been raised to 30 for experimentation.