Added a semi-automatic colour calibration
There are now automation options for changing the colour, controller and
tracker after setting the colour preset with a right-click. The automation
is prioritised as colour then controller then tracker. When cycling
through the colours the presets are automatically saved (the controller
and mouse must remain overlapped to be succesful).

Buttons have also been added to the colour configuration window to be able
to manually cycle through the controllers and trackers with out leaving
the window. A main menu button has also been added. The trailing zeros of
the frame rate, exposure and gain have been removed.

The buttons for changing the tracker in the tracker setting window are now
looped such that the first and last trackers can be switch to from one
another with a single click.