Modified the cross hairs shrink along with the size of the tracked orb
The cross hairs are now bound by the size of the marker they are drawn
within.

Changed the numbering of the trackers in the verification step of tracker
pose calibration because it was the only the had a "+1" so that it didn't
start at zero like normal.

Also added buttons to the test tracking video feeds to more easily switch
between them.