Added controller specific SteamVR button mapping
Added a section to steamvr.vrsettings for controller specific button
mapping based on the controller's Id (eg. add "_0" to the "psmove" section to
make it controller 0 specific). This also applies to the touchpad section
psmove_touchpad_directions. Coding the Id int to convert into a char was
problematic so it was done using if statements for values 0 to 9.

The controller Id parameter was added to the end of LoadButtonMapping with
a default value of -1 so that if it is not specified this new mapping
feature will be ignored. It is also ignored for Ids larger than 9.