Keep TriggerValue an unsigned char untill it is used
The TriggerValue for the PSMove and PSNavi were getting converted back
and forth between an unsigned char and a normalised float. To make it
simpler keep it as an unsigned char untill it needs to be read by a
program at which point it can be converted.

This will probably cause a problem for the PSMoveFreepieBridge but that
will need to be ungraded for the CAPI anyway.