Rotated acceleration vector for world view of test accelerometer
The acceleration vector was not properly set for the world view mode
of test accelerometer in the config tool. It had to be rotated by -90
deg (ie. oposite to how the defaultControllerTransform is set for the
PSMove controller model). The acceleration vector now always points up if
the controller is at rest in the world's frame of reference.