Fixed the serial number handling
Previously only one serial number was kept in the list (it wasn't an
actual list). This was the case for both the controller_serial list and
the parent_controller_serial list.

The PSNavi child check was also fixed. It was checking if it was a
PSMove instead of a PSNavi.