Add PSEye resolution control
Added control over the PSEye frame width and height. Since they are
linked only the width is specified in the configuration files but both
are available to be changed. They have been implemented similar to the
frame rate. The frame rate calls have been slightly changed so that they
match the frame width and height (ie. Framerate has been renamed FrameRate).

When changing the frame resolution the buffer is closed and reopened by
ServerTrackerView so that it has the correct size. This causes the video
feed in Color calibration to freeze so it is automatically exited and
reopened.

The QVGA resolution may be useful for the position triangulation method
if more than four cameras are used.