Fixes for PSMoveConfigTool
1) gotoColorCalib flag not getting reset to false after use

After changing the frame resolution the gotoColorCalib flag was
permanently set to true so nothing else on the trecker settings menu
could be used.

2) Controller color not set properly in color calibration

When switching the controllers the applied color was not set to the new
controller's default color.

3) Added images from the Color Calibration window

The images are to go into the wiki to explain how to use the semi auto
calibration, frame width and resolution options.