Added an optinal alignment marker to the colour calibration window
To help in aligning the trackers, an optional alignment marker has
been added to the Calibrate Controller Tracking Colours window. This
marker can be accessed by pressing the X key and has two modes (cycled
with the X key). The first mode uses a black cross to mark the absolute
center. The second mode uses the color oposite to the current orb color.