Integrate the alignment marker into the auto colour calibration
* Since the position of the orb is the same for all trackers when using
the alignment marker, this position can be used instead of the mouse
position. This also alows the calibration to continue uninterupted when
the tracker is changed. Auto controller switching however is then disabled.

* The surface used to place the controller on may not be vertically
centered (such as the floor) so the marker can now be moved up and down
using the W and S keys and recentered with the Z key.

* A similar marker has also been added to the verification step of the
tracker pose calibration but only has a black cross and it's vertical
position isn't synced with the other marker.